% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ephylo.R
\name{as.ephylo}
\alias{as.ephylo}
\title{Conversion among tree objects}
\usage{
as.ephylo(x, ...)
}
\arguments{
\item{x}{A \code{phylo} or \code{multiPhylo} object.}
}
\description{
\code{as.ephylo} is a generic function that converts \code{phylo} and
\code{multiPhylo} objects to \code{ephylo} objects.
}
