% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ephylo.R
\name{as.phylo.ephylo}
\alias{as.phylo.ephylo}
\title{S3 method for generic ape::as.phylo}
\usage{
\method{as.phylo}{ephylo}(x, ...)
}
\arguments{
\item{x}{An \code{ephylo} object.}

\item{...}{Additional arguments.}
}
\description{
S3 method for generic ape::as.phylo
}
