% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{C_ephylo}
\alias{C_ephylo}
\alias{C_ephylo_postorder}
\alias{C_ephylo_preorder}
\alias{C_ephylo_preorder_tips}
\alias{C_ephylo_read_newick}
\alias{C_mkcor_em}
\alias{C_mkcor_loglik}
\alias{C_mkcor_simulate}
\title{Call routine to internal C function}
\description{
Call routine to internal C function
}
