% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ephylo.R
\name{drop.tip.ephylo}
\alias{drop.tip.ephylo}
\title{S3 method for generic ape::drop.tip}
\usage{
\method{drop.tip}{ephylo}(phy, tip, ...)
}
\arguments{
\item{phy}{An \code{ephylo} object.}

\item{tip}{Integer vector with indices of tips to drop.}

\item{...}{Additional arguments.}
}
\description{
S3 method for generic ape::drop.tip
}
