% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ephylo.R
\name{ephylo}
\alias{ephylo}
\title{Extended \code{phylo} object}
\usage{
ephylo(phy)
}
\arguments{
\item{phy}{An object inheriting from class \code{phylo}.}
}
\value{
An S3 object with class attribute \code{c("ephylo","phylo")}. This
has the usual components and node indexing schema of a \code{phylo} object
 and adds the following additional components:
\describe{
  \item{parent}{An integer vector with the index of every node's parent.
    \code{parent[u]} returns the index of node \code{u}'s immediate ancestor
    or \code{0} if \code{u} is the root node.}
  \item{left.child}{An integer vector with the index of every node's left
    child. \code{left.child[u]} returns the index of node \code{u}'s left
    child or \code{0} if \code{u} has no left child.}
  \item{right.child}{An integer vector with the index of every node's right
    child. \code{right.child[u]} returns the index of node \code{u}'s right
    child or \code{0} if \code{u} has no right child.}
  \item{left.sib}{An integer vector with the index of every node's left
    sibling. \code{left.sib[u]} returns the index of node \code{u}'s left
    sibling or \code{0} if \code{u} has no left sibling.}
  \item{right.sib}{An integer vector with the index of every node's right
    sibling. \code{right.sib[u]} returns the index of node \code{u}'s right
    sibling or \code{0} if \code{u} has no right sibling.}
}
}
\description{
Turn a \code{phylo} object into a quintuply linked tree.
}
