% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ephylo.R
\name{ephylo_postorder}
\alias{ephylo_postorder}
\title{Postorder tree traversal}
\usage{
ephylo_postorder(phy, root)
}
\arguments{
\item{phy}{An object inheriting from class \code{ephylo}.}

\item{root}{Index of the node that will be the root of the traversal.}
}
\value{
An integer vector of node indices in postorder traversal
sequence from the given root.
}
\description{
Visit the nodes of a phylogeny in postorder traversal sequence.
}
