% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ephylo.R
\name{ephylo_preorder}
\alias{ephylo_preorder}
\title{Preorder tree traversal}
\usage{
ephylo_preorder(phy, root)
}
\arguments{
\item{phy}{An object inheriting from class \code{ephylo}.}

\item{root}{Index of the node that will be the root of the traversal.}
}
\value{
An integer vector of node indices in preorder traversal
sequence from the given root.
}
\description{
Visit the nodes of a phylogeny in preorder traversal sequence.
}
