% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ephylo.R
\name{ephylo_rclade}
\alias{ephylo_rclade}
\title{Sample random clades from a phylogeny}
\usage{
ephylo_rclade(phy, min.size, max.size, replace = FALSE, include.root = FALSE)
}
\arguments{
\item{phy}{An object inheriting from class \code{ephylo}.}

\item{min.size}{The minimum clade size to sample.}

\item{max.size}{The maximum clade size to sample.}

\item{replace}{Should sampling be conducted with replacement?}

\item{include.root}{Should the root clade be a candidate for sampling?}
}
\value{
A function to draw repeated random samples. It takes a single
parameter \code{n} specifying the number of random samples to take.
}
\description{
Sample random clades from a phylogeny
}
\details{
Sampling is conducted so that every size of clade (as measured by
the number of tips it contains) has an equal probability of being drawn.
This does not mean that you will see a uniform distribution of clade sizes
when drawing repeated random samples because there are still many more
small clades than large clades.
}
