% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ephylo.R
\name{ephylo_tips}
\alias{ephylo_tips}
\title{Return the tip indices descended from a given node}
\usage{
ephylo_tips(phy, root)
}
\arguments{
\item{phy}{An object inheriting from class \code{ephylo}.}

\item{root}{Index of the node that will be the root of the traversal.}
}
\value{
An integer vector of tip indices descended from the given root.
}
\description{
Return the tip indices descended from a given node
}
