% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ephylo.R
\name{keep.tip.ephylo}
\alias{keep.tip.ephylo}
\title{S3 method for generic ape::keep.tip}
\usage{
\method{keep.tip}{ephylo}(phy, tip, ...)
}
\arguments{
\item{phy}{An \code{ephylo} object.}

\item{tip}{Integer vector with indices of tips to keep.}

\item{...}{Additional arguments.}
}
\description{
S3 method for generic ape::keep.tip
}
