% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkcor.R
\name{mkcor_em}
\alias{mkcor_em}
\title{Fit the correlated Mk model by expectation-maximization}
\usage{
mkcor_em(x, y, phy)
}
\arguments{
\item{x}{A data.frame with character state data. One column should be
named 'tip.label' and another should be named 'state.id'. Character state
id's should be stored as factors. If a terminal node state is
ambiguous just list it in multiple rows with one row for each possible
state. If a terminal node state is unknown it can simply be omitted from the
data.frame.}

\item{y}{A data.frame for the second character formatted like \code{x}.}

\item{phy}{An \code{ephylo} object.}
}
\value{
A function that can be used to fit a model of correlated evolution
of \code{x} and \code{y} by expectation-maximization. The parameters to the
function are:
\itemize{
\item \code{par} A vector of initial parameter estimates. The first two values are
  rates of independent change in x and y, respectively. If given, the third 
  value is the rate of simultaneous change in x and y.
\item \code{maxiter} The maximum number of EM iterations to perform (default 100).
\item \code{tol} Stop the EM algorithm when the log likelihood improves by less
  than this amount between consecutive iterations (default 0.001).
}
The function returns a list with the following components:
\itemize{
\item \code{par} The maximum likelihood estimates of the overall rates of
  character evolution and character correlation.
\item \code{value} The log likelihood of the fitted model.
\item \code{event.counts} The expected number of different types of character state
  changes summed over all branches.
\item \code{state.probs} Ancestral state probabilities for each node in \code{phy}.
  E.g., \code{state.probs[i,j]} is the probability that node \eqn{i} is in state 
  \eqn{j}.
\item \code{branch.counts} An array with the expected number of different types of 
  character state changes on each branch. E.g., \code{branch.counts[i,j,u]} returns
  the expected number of evolutionary transitions from state \eqn{i} to state \eqn{j}
  on the branch leading to node \eqn{u}.
\item \code{logL} The log likelihood for each iteration of the EM algorithm. The
  final value corresponds to the parameter estimates stored in the other
  list components.
\item \code{states} A two column matrix enumerating the states in the model as
  combinations of the different states in \code{x} and \code{y}.
}
}
\description{
Returns a function that can be used to fit the correlated Mk model
to a set of observed character state data on a phylogeny using the 
expectation-maximization algorithm.
}
\seealso{
\code{\link{mkcor_fit_em}}
}
