% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkcor.R
\name{mkcor_fit}
\alias{mkcor_fit}
\title{Fit the correlated Mk model with L-BFGS-B}
\usage{
mkcor_fit(
  x,
  y,
  phy,
  correlated = TRUE,
  par_init = NULL,
  lower = 1e-05,
  upper = 1,
  lower_init = 0,
  upper_init = 0.001,
  ndeps = 0.3,
  num_fits = 1L
)
}
\arguments{
\item{x}{A data.frame with character state data. One column should be
named 'tip.label' and another should be named 'state.id'. Character state
id's should be stored as factors. If a terminal node state is
ambiguous just list it in multiple rows with one row for each possible
state. If a terminal node state is unknown it can simply be omitted from the
data.frame.}

\item{y}{A data.frame for the second character formatted like \code{x}.}

\item{phy}{An \code{ephylo} object.}

\item{correlated}{A boolean indicating if the correlated model (TRUE) or
the independent model (FALSE) should be fit.}

\item{par_init}{Optional initial parameter estimates to seed the algorithm.
If not \code{NULL}, any value for \code{correlated} and \code{num_fits} is
ignored and a single optimization is run until termination.}

\item{lower}{A lower bound for rate parameters (>= 0, default 1e-5).}

\item{upper}{An upper bound for rate parameters (default 1).}

\item{lower_init}{A lower bound for initial rate parameters (>= 0) used to
seed the fitting algorithm (default 0).}

\item{upper_init}{An upper bound for initial rate parameters used to seed
the fitting algorithm (default 0.001).}

\item{ndeps}{A scalar that multiplies the parameter estimates. The resulting
vector is then passed as the \code{ndeps} control parameter when calculating
the Hessian matrix. See \code{optim} for details.}

\item{num_fits}{Return the best (highest log likelihood) of \code{num_fits} 
models.}
}
\value{
A list with the following components:
\itemize{
\item \code{par} The maximum likelihood estimates of the overall rates of
  character evolution and character correlation with their 95% confidence
  intervals.
\item \code{value} The log likelihood of the best-fitting model.
\item \code{counts} The number of likelihood evaluations made by \code{optim}.
\item \code{convergence} The convergence code returned by \code{optim}.
\item \code{message} Additional information returned by \code{optim}.
}
}
\description{
Return parameter estimates of the correlated Mk model given a set of 
observed character state data using the L-BFGS-B numerical optimization
algorithm.
}
\seealso{
The \code{vignette("CorrelatedMk")}
}
