% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkcor.R
\name{mkcor_fit_em}
\alias{mkcor_fit_em}
\title{Fit the correlated Mk model by expectation-maximization}
\usage{
mkcor_fit_em(
  x,
  y,
  phy,
  correlated = TRUE,
  par_init = NULL,
  maxiter = 100L,
  tol = 0.001,
  lower_init = 0,
  upper_init = 0.001,
  ndeps = 0.3,
  num_fits = 1L
)
}
\arguments{
\item{x}{A data.frame with character state data. One column should be
named 'tip.label' and another should be named 'state.id'. Character state
id's should be stored as factors. If a terminal node state is
ambiguous just list it in multiple rows with one row for each possible
state. If a terminal node state is unknown it can simply be omitted from the
data.frame.}

\item{y}{A data.frame for the second character formatted like \code{x}.}

\item{phy}{An \code{ephylo} object.}

\item{correlated}{A boolean indicating if the correlated model (TRUE) or
the independent model (FALSE) should be fit.}

\item{par_init}{Optional initial parameter estimates to seed the algorithm.
If not \code{NULL}, any value for \code{correlated} and \code{num_fits} is
ignored and a single EM chain is run until termination.}

\item{maxiter}{The maximum number of EM iterations to perform (default 100).}

\item{tol}{Stop the EM algorithm when the log likelihood improves by less
than this amount between consecutive iterations (default 0.001).}

\item{lower_init}{A lower bound for initial rate parameters (>= 0) used to
seed the fitting algorithm (default 0).}

\item{upper_init}{An upper bound for initial rate parameters used to seed
the fitting algorithm (default 0.001).}

\item{ndeps}{A scalar that multiplies the parameter estimates. The resulting
vector is then passed as the \code{ndeps} control parameter when calculating
the Hessian matrix. See \code{optim} for details. The Hessian is
used for calculating standard errors of parameter estimates.}

\item{num_fits}{Return the best (highest log likelihood) of \code{num_fits} 
models.}
}
\value{
A list with the following components:
\itemize{
\item \code{par} The maximum likelihood estimates of the overall rates of
  character evolution and character correlation along with their standard
  errors.
\item \code{value} The log likelihood of the best-fitting model.
\item \code{event.counts} The expected number of different types of character state
  changes summed over all branches.
\item \code{state.probs} Ancestral state probabilities for each node in \code{phy}.
  E.g., \code{state.probs[i,j]} is the probability that node \eqn{i} is in state 
  \eqn{j}.
\item \code{branch.counts} An array with the expected number of different types of 
  character state changes on each branch. E.g., \code{branch.counts[i,j,u]} returns
  the expected number of evolutionary transitions from state \eqn{i} to state \eqn{j}
  on the branch leading to node \eqn{u}.
\item \code{logL} The log likelihood for each iteration of the EM algorithm. The
  final value corresponds to the parameter estimates stored in the other
  list components.
\item \code{states} A two column matrix enumerating the states in the model as
  combinations of the different states in \code{x} and \code{y}.
}
}
\description{
Return parameter estimates of the correlated Mk model given a set of 
observed character state data using expectation-maximization.
}
\seealso{
The \code{vignette("CorrelatedMk")}
}
