% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkcor.R
\name{mkcor_loglik}
\alias{mkcor_loglik}
\title{Calculate the log likelihood of the correlated Mk model}
\usage{
mkcor_loglik(x, y, phy)
}
\arguments{
\item{x}{A data.frame with character state data. One column should be
named 'tip.label' and another should be named 'state.id'. Character state
id's should be stored as factors. If a terminal node state is
ambiguous just list it in multiple rows with one row for each possible
state. If a terminal node state is unknown it can simply be omitted from the
data.frame.}

\item{y}{A data.frame for the second character formatted like \code{x}.}

\item{phy}{An \code{ephylo} object.}
}
\value{
A function that calculates the log likelihood of parameters of
the correlated Mk model given the observed data. The function takes a
single argument:
\itemize{
\item \code{par} 
  A vector of transition rate parameters. The first two values are
  rates of independent change in x and y, respectively. If given, the third 
  value is the rate of simultaneous change in x and y.
}
}
\description{
Returns a function that calculates the log likelihood of parameters of
the correlated Mk model given a set of observed character state data.
}
\seealso{
\code{\link{mkcor_fit}}
}
