% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newick.R
\name{read_newick}
\alias{read_newick}
\title{Read a tree from a Newick file or string}
\usage{
read_newick(file, text, num_tips)
}
\arguments{
\item{file}{The path to the Newick file.}

\item{text}{The Newick string. Ignored if \code{file} is specified.}

\item{num_tips}{The number of tips in the phylogeny (optional). 
Specifying the number of tips in advance will speed the parsing
of large trees by eliminating the need for memory reallocations.}
}
\value{
An S3 object with class attribute \code{c("ephylo", "phylo")}.
}
\description{
Read a tree from a Newick file or string
}
\seealso{
\code{\link{ephylo}}
}
