% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ephylo.R
\name{reorder.ephylo}
\alias{reorder.ephylo}
\title{S3 method for generic stats::reorder that hooks into ape::reorder.phylo}
\usage{
\method{reorder}{ephylo}(x, ...)
}
\arguments{
\item{x}{An \code{ephylo} object.}

\item{...}{Additional arguments.}
}
\description{
S3 method for generic stats::reorder that hooks into ape::reorder.phylo
}
