% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkcor.R
\name{simulate_mkcor}
\alias{simulate_mkcor}
\title{Simulate data under the correlated Mk model}
\usage{
simulate_mkcor(par, num_states_x, num_states_y, phy)
}
\arguments{
\item{par}{A vector of transition rate parameters. The first two values are
rates of independent change in X and Y, respectively. The third 
value, if given, is the rate of simultaneous change in X and Y.}

\item{num_states_x}{The number of states in X.}

\item{num_states_y}{The number of states in Y.}

\item{phy}{An \code{ephylo} object.}
}
\value{
A list with named components `x' and `y', which are data.frame
objects containing the simulated tip states. Each data.frame has a
`node.state' attribute containing the simulated states at internal
nodes.
}
\description{
Simulate data under the correlated Mk model
}
