% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{squamate_dental_diet_states}
\alias{squamate_dental_diet_states}
\title{Squamate reptile dental-diet states from Lafuma et al. (2021)}
\format{
An object of class \code{data.frame} with 548 rows and 18 columns.
}
\source{
Lafuma, F., Corfe, I.J., Clavel, J., and Di-Poï, N. 2021.
  Multiple evolutionary origins and losses of tooth complexity in squamates.
  Nature Communications 12(1). doi: 10.1038/s41467-021-26285-w
}
\usage{
squamate_dental_diet_states
}
\description{
A \code{data.frame} of dental-diet states
}
\keyword{datasets}
